manage listings

+++ app/Http/Controllers/ListingController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class ListingController extends Controller
@@ -41,15 +43,32 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
         return redirect('/')->with('message','Listing created successfully!');
     }
     //Edit Listing
+
     public function edit(Listing $listing){
+        //Check Authenticated User is Owner
+        $userid = $listing->user_id;
+        $authid = auth()->id();
+        if($userid != $authid){
+            return view('error');
+        }
         return view('listings.edit',['listing'=>$listing]);
     }
+
     //update listing
     public function update(Request $request, Listing $listing){
+
+        //Check Authenticated User is Owner
+        $userid = $listing->user_id;
+        $authid = auth()->id();
+        if($userid != $authid){
+            abort(403, 'Unauthorized action baby');
+        }
 
         $formFields = $request->validate([
             'title'         =>  'required',
@@ -70,7 +89,21 @@ class ListingController extends Controller
     }
     //Delete Listing
     public function destroy(Listing $listing){
+        
+        //Check Authenticated User is Owner
+        $userid = $listing->user_id;
+        $authid = auth()->id();
+        if($userid != $authid){
+            return view('error');
+        }
         $listing->delete();
         return redirect('/')->with('message','Listing was deleted!');
+    }
+
+    //Manage listings
+    public function manage(){
+        return view('listings.manage',[
+            'listings' => auth()->user()->listings
+        ]);
     }
 }
